+++ subcommands/createVersion/index.js
@@ -51,6 +51,7 @@ const createVersion = async (bkName, cmdOptions) => {
       spinnies.add('sync', { text: 'Checking sync status' })
       await tempSync(null, { returnOnError: true })
       spinnies.succeed('sync', { text: 'sync is up to date' })
+      console.log();
 
       if (!existsSync(orphanBranchFolder)) throw new Error(`Error reading bb modules block_${blockName}`)
       if (!existsSync(workSpaceFolder)) throw new Error(`Error reading bb modules workspace`)
@@ -97,7 +98,7 @@ const createVersion = async (bkName, cmdOptions) => {
 
       // called from package context
 
-      if (!blockName || blockName === manager.config.blockName) {
+      if (!blockName || blockName === manager.config.name) {
         // if blockName is empty or passed blockName is of package
         packageManager = manager
       } else {
@@ -117,14 +118,14 @@ const createVersion = async (bkName, cmdOptions) => {
       if (manager instanceof BlockConfigManager) {
         // called from block context
 
-        if (blockName && blockName !== manager.config.blockName) {
+        if (blockName && blockName !== manager.config.name) {
           throw new Error('Cannot pass block name inside block context')
         }
         blockManager = manager
       } else if (manager instanceof PackageConfigManager) {
         // called from package context
 
-        if (!blockName || blockName === manager.config.blockName) {
+        if (!blockName || blockName === manager.config.name) {
           // if passed blockName is of package
           packageManager = manager
         } else {