+++ subcommands/init/setupTemplateV2.js
@@ -1,13 +1,13 @@
 const { cp, readFile, writeFile } = require('fs/promises')
 const path = require('path')
 const { nanoid } = require('nanoid')
-const { existsSync } = require('fs')
+const { existsSync, readdir } = require('fs')
 const { BB_CONFIG_NAME } = require('../../utils/constants')
 
 async function setupTemplateV2(options) {
-  const { DIRPATH, blockVisibility, packageBlockId, packageParentBlockIDs, repoType,packagename} = options
+  const { DIR_PATH, blockVisibility, packageBlockId, packageParentBlockIDs, repoType, packageName } = options
 
-  const configPath = path.join(DIRPATH, BB_CONFIG_NAME)
+  const configPath = path.join(DIR_PATH, BB_CONFIG_NAME)
   const config = await readFile(configPath, { encoding: 'utf8' })
 
   const templatesPath = path.join(__dirname, '..', '..', 'templates', 'sample-todo-template-v2')
@@ -24,27 +24,38 @@ async function setupTemplateV2(options) {
   b.parentBlockIDs = packageParentBlockIDs
   b.source.branch = `block_${b.name}`
 
-  await cp(templatesPath, DIRPATH, { recursive: true })
+  readdir(templatesPath, { withFileTypes: true }, async (err, files) => {
+    if (err) throw err
 
-  Promise.allSettled(
-    Object.keys(b.dependencies).map(async (blockName) => {
-      const blockConfigPath = path.join(DIRPATH, b.dependencies[blockName].directory, BB_CONFIG_NAME)
-      const currentBlock = JSON.parse(await readFile(blockConfigPath, 'utf8'))
-      if (existsSync(blockConfigPath)) {
-        currentBlock.name=`${packagename}_${currentBlock.name}`
-        currentBlock.blockId = nanoid()
-        currentBlock.isPublic = blockVisibility
-        currentBlock.repoType = repoType
-        currentBlock.parentBlockIDs = [...packageParentBlockIDs, b.blockId]
-        currentBlock.source.branch = `block_${currentBlock.name}`
+    // Filter templateBlock directories
+    const templateBlocks = files.filter((file) => file.isDirectory()).map((file) => file.name)
+    const newDependencies = {}
+    await Promise.allSettled(
+      templateBlocks.map(async (templateBlockName) => {
+        const newBlockName = `${packageName}_${templateBlockName}`
+        const newBlockPath = path.join(DIR_PATH, newBlockName)
+        await cp(path.join(templatesPath, templateBlockName), newBlockPath, { recursive: true })
 
-        await writeFile(blockConfigPath, JSON.stringify(currentBlock, null, 2))
-      }
-    })
-  )
+        newDependencies[newBlockName] = { directory: newBlockName }
+        const blockConfigPath = path.join(newBlockPath, BB_CONFIG_NAME)
+        if (existsSync(blockConfigPath)) {
+          const currentBlock = JSON.parse(await readFile(blockConfigPath, 'utf8'))
 
-  await writeFile(path.join(DIRPATH, BB_CONFIG_NAME), JSON.stringify(b, null, 2))
-  // await writeFile(path.join(DIRPATH, '.gitignore'), 'node_modules\n')
+          currentBlock.name = `${packageName}_${currentBlock.name}`
+          currentBlock.blockId = nanoid()
+          currentBlock.isPublic = blockVisibility
+          currentBlock.repoType = repoType
+          currentBlock.parentBlockIDs = [...packageParentBlockIDs, b.blockId]
+          currentBlock.source.branch = `block_${currentBlock.name}`
+
+          await writeFile(blockConfigPath, JSON.stringify(currentBlock, null, 2))
+        }
+      })
+    )
+
+    b.dependencies = newDependencies
+    await writeFile(path.join(DIR_PATH, BB_CONFIG_NAME), JSON.stringify(b, null, 2))
+  })
 }
 
 module.exports = setupTemplateV2