+++ subcommands/tempSync/createBBModules/util.js
@@ -272,7 +272,7 @@ const setVisibilityAndDefaultBranch = async (options) => {
   const { configstore, repoUrl, headLessConfigStore } = options
 
   let defaultBranch = headLessConfigStore.get('defaultBranch')
-  // let repoVisibility = headLessConfigStore.get('repoVisibility')
+  // let repoVisibility = headLessConfigStore().get('repoVisibility')
 
   if (defaultBranch) {
     return { defaultBranch }
@@ -320,7 +320,7 @@ const setVisibilityAndDefaultBranch = async (options) => {
   //   repoVisibility = inputRepoVisibility
   // }
 
-  // headLessConfigStore.set('repoVisibility', repoVisibility)
+  // headLessConfigStore().set('repoVisibility', repoVisibility)
 
   if (defaultBranch.length === 0) {
     console.log('entered repo main branch unavailable\n')